requirements.txt
attr==0.3.2
beautifulsoup4==4.12.3
ConfigParser==6.0.1
contextlib2==21.6.0
cryptography==42.0.5
customtkinter==5.2.2
datasets==2.18.0
docutils==0.20.1
filelock==3.13.1
HTMLParser==0.0.2
importlib_metadata==7.1.0
ipython==8.12.3
ipywidgets==8.1.2
Jinja2==3.1.3
jnius==1.1.0
jobspy==0.29.0
keyring==24.3.1
mock==5.1.0
nltk==3.8.1
openai==1.14.2
pandas==2.2.1
pdfminer==20191125
pdfquery==0.4.3
Pillow==10.2.0
protobuf==5.26.0
pyodbc==5.1.0
pyOpenSSL==24.1.0
python_dateutil==2.9.0.post0
railroad==0.5.0
redis==5.0.3
scikit_learn==1.4.1.post1
spacy==3.7.4
Sphinx==7.2.6
thread==1.1.1
tornado==6.4
tqdm==4.66.2
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
geocoder==1.38.1